number refactor math.comb
