+++ src/632.py
@@ -11,8 +11,8 @@
 # This is the generalized inclusion-exclusion principle and the coefficient for
 # the intersection of m sets is (-1)^(m+k) C(m, k)
 
-from math import prod
-from number import sieve, comb
+from math import prod, comb
+from number import sieve
 
 def C(N):
     sqrt_N = int(N**0.5)
